combine webm demuxers into a base class
